% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{analyse_data}
\alias{analyse_data}
\title{Reads the necessary data and defines variables used in later functions}
\usage{
analyse_data(profs_data)
}
\description{
Reads the necessary data and defines variables used in later functions
}
\examples{
data(allprofs)
analyse_data(allprofs)
}

