% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean.R
\name{get_mean}
\alias{get_mean}
\title{Prints the mean age in each division}
\usage{
get_mean()
}
\description{
Prints the mean age in each division
}
\examples{
library(ggplot2)
data(allprofs)
analyse_data(allprofs)
get_mean()
}

