% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot1.R
\name{get_plot1}
\alias{get_plot1}
\title{Prints a histogram with the divisions that the professors belong to}
\usage{
get_plot1()
}
\description{
Prints a histogram with the divisions that the professors belong to
}
\examples{
library(ggplot2)
data(allprofs)
analyse_data(allprofs)
get_plot1()
}

