% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot3.R
\name{get_plot3}
\alias{get_plot3}
\title{Prints a scatter plot of ages and the year the professor received the last degree}
\usage{
get_plot3()
}
\description{
Prints a scatter plot of ages and the year the professor received the last degree
}
\examples{
library(ggplot2)
data(allprofs)
analyse_data(allprofs)
get_plot3()
}

