% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Prints out a table of all professors, ordered by age}
\usage{
get_table()
}
\description{
Prints out a table of all professors, ordered by age
}
\examples{
data(allprofs)
analyse_data(allprofs)
get_table()
}

