% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_data}
\alias{read_data}
\title{Reads the necessary data and defines variables used in later functions}
\usage{
read_data()
}
\description{
Reads the necessary data and defines variables used in later functions
}
\examples{
read_data()
}

