% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyse_data}
\alias{analyse_data}
\title{Reads the necessary data and defines variables used in later functions}
\usage{
analyse_data(profs_data)
}
\description{
Reads the necessary data and defines variables used in later functions
}
\examples{
read_data()
}

