% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_plot1}
\alias{get_plot1}
\title{Prints a histogram with the divisions that the professors belong to}
\usage{
get_plot1()
}
\description{
Prints a histogram with the divisions that the professors belong to
}
\examples{
get_plot1()
}

