% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot2.R
\name{get_plot2}
\alias{get_plot2}
\title{Prints a graph showing age distributions in each department}
\usage{
get_plot2()
}
\description{
Prints a graph showing age distributions in each department
}
\examples{
library
data(allprofs)
analyse_data(allprofs)
get_plot2()
}

